Add files via upload
Cybertrigo presents a  automated command-line interface to VirusTotal, Jotti, and NoVirusThanks.
The script gives you the ability to analyze files using multiple services without using a web
browser or a special client. Since it is written in Python, it works on Linux^, Mac OS X^, and
Windows^. You must not use this script for commercial purposes or in manner that violates
the vendor’s acceptable use policy.
With the ability to upload files on the Ctaviav command line, you can easily automate submissions
and retrieve the results. For example, you could create a second script to extract
potentially dangerous attachments from a local MBOX file or from a remote POP3/IMAP
account; then pass the attachments to Ctaviav.py